// src/app/shared/models/nyilak.model.ts
export interface Nyilak {
  /**
   * van 1-5-9
   */
  eltokeltseg: boolean;

  /**
   * nincs 1-5-9
   */
  halogatas: boolean;

  /**
   * van 3-5-7
   */
  spiritualitas: boolean;

  /**
   * nincs 3-5-7
   */
  szkepticizmus: boolean;

  /**
   * van 3-6-9
   */
  intellektus: boolean;

  /**
   * nincs 3-6-9
   */
  rosszMemoria: boolean;

  /**
   * van 2-5-8
   */
  erzelmiKiegyensulyozottsag: boolean;

  /**
   * nincs 2-5-8
   */
  hipererzekenyseg: boolean;

  /**
   * van 1-4-7
   */
  gyakorlatiassag: boolean;

  /**
   * nincs 1-4-7
   */
  rendezetlenseg: boolean;

  /**
   * van 3-2-1
   */
  tervezes: boolean;

  /**
   * van 6-5-4
   */
  akarat: boolean;

  /**
   * nincs 4-5-6
   */
  csalodottsag: boolean;

  /**
   * van 7-8-9
   */
  aktivitas: boolean;

  /**
   * nincs 7-8-9
   */
  passzivitas: boolean;
}